Suppress INFO level logs from phantomjs
Running integration tests makes phantomjs pollute the log with unnecessary
information. Fix that by:
  1. creating a PhantomJS test browser that can be given the argument to
     silence INFO level logs,
  2. giving JVM necessary logging option.

In addition define the organization name of the project (based on github
account), add test:grunt task back to fullTest, and change scalastyle rules
a little (intention was to enable ScalaDoc checker in scalastyle but it's
disabled because of it giving false positives).

Closes #2.